+++ CHANGELOG.md
@@ -35,7 +35,7 @@ Unreleased
 * Improved handling of branch and gen outages: Outages can be enabled by setting the outage flag. Contingencies no longer disconnect components.
 * Added bus.is_star(), branch.is_part_of_3_winding_transformer(), and net.get_num_star_buses().
 * Added routines for getting number of gens/branches on outage.
-* Added utility routines to check function gradients in pfnet.tests.utils.
+* Added utility routines to check function gradients and hessians in pfnet.tests.utils.
 
 Version 1.3.2
 -------------